walking_detection/cloud_computing/azure_function_cloud/requirements.txt
azure-functions
pyodbc
datetime
requests